Proiect/CloudProjectClient-master/datastore_functions/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.8.0